# auth-service/openapi.yaml
openapi: 3.0.0
paths:
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: ''
      tags: &ref_0
        - Auth
  /auth/verify:
    post:
      operationId: AuthController_verify
      parameters:
        - name: token
          required: true
          in: query
          schema:
            type: string
      responses:
        '201':
          description: ''
      tags: *ref_0
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '201':
          description: ''
      tags: *ref_0
  /auth/forgot:
    post:
      operationId: AuthController_forgot
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '201':
          description: ''
      tags: *ref_0
  /auth/reset:
    post:
      operationId: AuthController_reset
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '201':
          description: ''
      tags: *ref_0
...
  /tenants/{tid}/roles/{rid}:
    get:
      operationId: RoleManagementController_findOne
      parameters:
        - name: tid
          required: true
          in: path
          schema:
            type: number
        - name: rid
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_2
    patch:
      operationId: RoleManagementController_update
      parameters:
        - name: tid
          required: true
          in: path
          schema:
            type: number
        - name: rid
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
    delete:
      operationId: RoleManagementController_remove
      parameters:
        - name: tid
          required: true
          in: path
          schema:
            type: number
        - name: rid
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
      tags: *ref_2
...
info:
  title: My App API
  description: API docs for My App
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      description: Enter JWT token
      in: header
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: john.doe@example.com
        password:
          type: string
          example: Aman12345
        username:
          type: string
          example: AmanT16
        full_name:
          type: string
          example: John Doe
        tenantId:
          type: number
          example: 1
      required:
        - email
        - password
        - username
        - full_name
        - tenantId
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: JOhndoe1234@gmail.com
        password:
          type: string
          example: Aman12345
      required:
        - email
        - password
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          example: JOhndoe1234@gmail.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        newPassword:
          type: string
          example: JOhndoe1234
      required:
        - newPassword
...
    UpdateRoleDto:
      type: object
      properties: {}
...